refactor: `/components` -> `/app/components`
