+++ examples/with-stripe-typescript/app/components/CustomDonationInput.tsx
@@ -1,5 +1,4 @@
-import React from 'react'
-import { formatAmountForDisplay } from '../utils/stripe-helpers'
+import { formatAmountForDisplay } from '../../utils/stripe-helpers'
 
 type Props = {
   name: string
