+++ examples/with-stripe-typescript/app/donate-with-elements/result/page.tsx
@@ -1,6 +1,6 @@
 import Stripe from 'stripe'
 
-import PrintObject from '../../../components/PrintObject'
+import PrintObject from '../../components/PrintObject'
 
 const stripe = new Stripe(process.env.STRIPE_SECRET_KEY!, {
   // https://github.com/stripe/stripe-node#configuration
